Configurable Task module
`Task.async` doesn't preserve metadata that's stored in the process
dictionary. Known impacts for this are:

* Logger metadata is not preserved. This means log entries from an async
  task will not include this extra information, such as request ID.

* Tracing tools such as Spandex may not be able to provide continuity
  into async tasks. Tracing information within them is lost.

One way to solve the above issues is with a function that wraps
`Task.async`; however, to extend this solution to dependencies that use
`Task.async`, the dependency must have a way to accept the wrapper
function.

This commit introduces a configuration option, `task_module`, which allows
applications to configure an alternative module to use for `Task.async`
and `Task.await`.